refactor: test the new helpers from the external test packages
The repo has no _internal_test.go files and tests command helpers through
the external test package, exporting what they need to reach. Follow that:
export ResolveTaskIDs and ShouldSpin, fold the task ID cases into the
existing wait_test.go, and move the round-tripper tests to apiclient_test
alongside the other two test files in that package.
